Refactor tests, improve CLI menu, and update compiler flags.
Replaced `int` with `size_t` in tests for safer iteration and fixed a linter warning in `VerifyPassword`. Improved CLI menu by moving the explicit exit option and updating output formatting. Enabled stricter compiler warnings for better code quality.

+++ tests/src/tests.cpp
@@ -189,7 +189,7 @@ TEST_F(PasswordGenerationTests1, HashPasswordsSafeAsyncCorrectlyHashes)
 
     // then:
     const auto encrypted = encryptedFut.get();
-    for (int i = 0; i < passwords.size(); i++)
+    for (size_t i = 0; i < passwords.size(); i++)
     {
         EXPECT_TRUE(passwordGenerator.VerifyPasswordSafe(passwords[i], encrypted[i])) << "Password hash verification failed";
     }
